+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.4.4] - NOT RELEASED
 ##### Added
 - Added improved loading indicators for net worth page
+- Minor visual improvements
 ##### Fixed
 - Fixed a bug where the tabs column would take up multiple rows