+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.4.4] - NOT RELEASED
+## [0.4.4] - 2021-03-30
 ##### Added
 - Added improved loading indicators for net worth page
 - Minor visual improvements